requirements.txt
Django==1.4.3
https://github.com/zorna/django-schedule/archive/master.zip
git+git://github.com/justquick/django-activity-stream.git#egg=django-activity-stream
BeautifulSoup==3.2.1
PIL==1.1.7
Pillow==1.7.8
PyYAML==3.10
South==0.7.6
Whoosh==2.4.1
django-ckeditor==3.6.2.1
django-haystack==1.2.7
django-mptt==0.5.5
django-tagging==0.3.1
feedparser==5.1.3
python-dateutil==2.1
six==1.1.0
vobject==0.8.1c
# MySQL-python==1.2.4

